♻️ Extract main Supervisor to its own module
Separates the responsibility of starting/supervising the app from the public API.

Add a delegator `child_spec` function to `ConfigCat` so that client apps can continue to include `ConfigCat` in their supervision trees.
